+++ src/store/modules/index.js
@@ -1,7 +1,9 @@
 import Page from '@/store/modules/page'
+import Pages from '@/store/modules/pages'
 import User from '@/store/modules/user'
 
 export default {
   Page,
+  Pages,
   User,
 }