+++ src/store/mutation-types.js
@@ -1,3 +1,4 @@
 export const PAGE_UPDATE_URL = 'UPDATE_URL'
 export const PAGE_UPDATE_CONTENT = 'UPDATE_CONTENT'
+export const PAGES_UPDATE = 'PAGES_UPDATE'
 export const USER_LOGIN = 'USER_LOGIN'